Made a further project restructuring
