+++ page_ranker_app/source/page_ranker.py
@@ -1,9 +1,8 @@
 import re
+import requests
 import threading
 import time
 import timeit
-
-import requests
 import urllib.parse
 
 from collections import defaultdict
@@ -12,7 +11,7 @@ from itertools import repeat
 from typing import List
 
 from page_ranker_app import settings
-from page_ranker_app.source import crawlers, inverters, parsers, utils
+from page_ranker_app.source import crawlers, inverters, parsers
 
 
 class PageRankInfoAccumulator:
@@ -127,27 +126,5 @@ class WikiPageRankInfoAccumulator(PageRankInfoAccumulator):
         self._page_rank = {key: len(value) for key, value in rev_data.items()}
 
 
-def main(url: str, limit: int):
-    wiki_scrapper = WikiPageRankInfoAccumulator(url, limit)
-    with utils.timer():
-        wiki_scrapper.scrap_data_till_limit()
-    # print(wiki_scrapper._page_links)
-    with utils.timer():
-        wiki_scrapper.count_page_rank()
-        dict1 = wiki_scrapper.page_rank
-    # print(wiki_scrapper.page_rank)
-    with utils.timer():
-        wiki_scrapper.inverter = inverters.DictionaryInverterProcessing
-        wiki_scrapper.count_page_rank()
-        dict2 = wiki_scrapper.page_rank
-    with utils.timer():
-        wiki_scrapper.inverter = inverters.DictionaryInverterSync
-        wiki_scrapper.count_page_rank()
-        dict3 = wiki_scrapper.page_rank
-    print(dict1 == dict2 == dict3)
-
-
 if __name__ == "__main__":
-    test_url = "https://en.wikipedia.org/wiki/Superintendent"
-    test_limit = 100
-    main(test_url, test_limit)
+    pass
